relocate emojiAutoCompletionSettings.ts,
